+++ packages/sdk/src/models/field.ts
@@ -281,6 +281,83 @@ class Field extends AbstractModel<FieldData, WatchableFieldKey> {
     }
 
     /**
+     * Checks whether the current user has permission to perform the given name update.
+     *
+     * Accepts partial input, in the same format as {@link updateNameAsync}.
+     *
+     * Returns `{hasPermission: true}` if the current user can update the specified field,
+     * `{hasPermission: false, reasonDisplayString: string}` otherwise. `reasonDisplayString` may be
+     * used to display an error message to the user.
+     *
+     * @param name new name for the field
+     *
+     * @example
+     * ```js
+     * const updateFieldCheckResult = field.checkPermissionsForUpdateName();
+     *
+     * if (!updateFieldCheckResult.hasPermission) {
+     *     alert(updateFieldCheckResult.reasonDisplayString);
+     * }
+     * ```
+     */
+    checkPermissionsForUpdateName(name?: string): PermissionCheckResult {
+        return this._sdk.__mutations.checkPermissionsForMutation({
+            type: MutationTypes.UPDATE_SINGLE_FIELD_NAME,
+            tableId: this.parentTable.id,
+            id: this.id,
+            name,
+        });
+    }
+
+    /**
+     * An alias for `checkPermissionsForUpdateName(options).hasPermission`.
+     *
+     * Checks whether the current user has permission to perform the name update.
+     *
+     * Accepts partial input, in the same format as {@link updateNameAsync}.
+     *
+     * @param name new name for the field
+     *
+     * @example
+     * ```js
+     * const canUpdateField = field.hasPermissionToUpdateName();
+     *
+     * if (!canUpdateField) {
+     *     alert('not allowed!');
+     * }
+     * ```
+     */
+    hasPermissionToUpdateName(name?: string): boolean {
+        return this.checkPermissionsForUpdateName(name).hasPermission;
+    }
+
+    /**
+     * Updates the name for this field.
+     *
+     * Throws an error if the user does not have permission to update the field, or if an invalid
+     * name is provided.
+     *
+     * This action is asynchronous. Unlike updates to cell values, updates to field name are
+     * **not** applied optimistically locally. You must `await` the returned promise before
+     * relying on the change in your app.
+     *
+     * @param name new name for the field
+     *
+     * @example
+     * ```js
+     * await myTextField.updateNameAsync('My New Name');
+     * ```
+     */
+    async updateNameAsync(name: string): Promise<void> {
+        await this._sdk.__mutations.applyMutationAsync({
+            type: MutationTypes.UPDATE_SINGLE_FIELD_NAME,
+            tableId: this.parentTable.id,
+            id: this.id,
+            name,
+        });
+    }
+
+    /**
      * Checks whether the current user has permission to perform the given description update.
      *
      * Accepts partial input, in the same format as {@link updateDescriptionAsync}.